requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
lxml==4.6.2
numpy==1.19.5
openpyxl==3.0.6
pandas==1.1.5
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
six==1.15.0
xlrd==2.0.1
